+++ CHANGELOG.md
@@ -11,6 +11,7 @@ with exception of a "0." prefix until version 1 is released.
 
  - Type `files-matcher` - complex expressions (&& and ||)
  - Type `string-matcher` - complex expressions (&& and ||)
+ - Type `file-matcher` - primitive `dir-contents`
 
 ## [0.11.1.0]